move mockSearcher to internal package
